Updates to carnivore.json
* Removed a single remaining dummied-out item that was inexplicably
causing a "type not found" error with no indication of the problem file.
Dunno why, when all the other dummied out cannibal food gave errors that
made the cause easier to find.
* Also added `copy-from`s for most of the carnivore items that've since
been added. Weird because they SHOULD be getting freeze points from your
edits to the ones they copy from, but for some crazy reason only a few
of the offending items are doing that properly.